Shrink Node size by making block numbers and indicies uint64s

+++ solgen/src/rollup/IRollupCore.sol
@@ -27,8 +27,8 @@ interface IRollupCore {
         external
         view
         returns (
-            uint256,
-            uint256,
+            uint64,
+            uint64,
             uint256,
             IChallenge,
             bool
@@ -48,18 +48,18 @@ interface IRollupCore {
     );
 
     event NodeConfirmed(
-        uint256 indexed nodeNum,
+        uint64 indexed nodeNum,
         bytes32 blockHash,
         bytes32 sendRoot
     );
 
-    event NodeRejected(uint256 indexed nodeNum);
+    event NodeRejected(uint64 indexed nodeNum);
 
     event RollupChallengeStarted(
         IChallenge indexed challengeContract,
         address asserter,
         address challenger,
-        uint256 challengedNode
+        uint64 challengedNode
     );
 
     event UserStakeUpdated(address indexed user, uint256 initialBalance, uint256 finalBalance);
@@ -73,19 +73,19 @@ interface IRollupCore {
     /**
      * @notice Get the Node for the given index.
      */
-    function getNode(uint256 nodeNum) external view returns (Node memory);
+    function getNode(uint64 nodeNum) external view returns (Node memory);
 
     /**
      * @notice Check if the specified node has been staked on by the provided staker
      */
-    function nodeHasStaker(uint256 nodeNum, address staker) external view returns (bool);
+    function nodeHasStaker(uint64 nodeNum, address staker) external view returns (bool);
 
     /**
      * @notice Get the address of the staker at the given index
      * @param stakerNum Index of the staker
      * @return Address of the staker
      */
-    function getStakerAddress(uint256 stakerNum) external view returns (address);
+    function getStakerAddress(uint64 stakerNum) external view returns (address);
 
     /**
      * @notice Check whether the given staker is staked
@@ -99,7 +99,7 @@ interface IRollupCore {
      * @param staker Staker address to lookup
      * @return Latest node staked of the staker
      */
-    function latestStakedNode(address staker) external view returns (uint256);
+    function latestStakedNode(address staker) external view returns (uint64);
 
     /**
      * @notice Get the current challenge of the given staker
@@ -127,7 +127,7 @@ interface IRollupCore {
      * @param zombieNum Index of the zombie to lookup
      * @return Latest node that the given zombie is staked on
      */
-    function zombieLatestStakedNode(uint256 zombieNum) external view returns (uint256);
+    function zombieLatestStakedNode(uint256 zombieNum) external view returns (uint64);
 
     /// @return Current number of un-removed zombies
     function zombieCount() external view returns (uint256);
@@ -145,20 +145,20 @@ interface IRollupCore {
      * @return Index of the first unresolved node
      * @dev If all nodes have been resolved, this will be latestNodeCreated + 1
      */
-    function firstUnresolvedNode() external view returns (uint256);
+    function firstUnresolvedNode() external view returns (uint64);
 
     /// @return Index of the latest confirmed node
-    function latestConfirmed() external view returns (uint256);
+    function latestConfirmed() external view returns (uint64);
 
     /// @return Index of the latest rollup node created
-    function latestNodeCreated() external view returns (uint256);
+    function latestNodeCreated() external view returns (uint64);
 
     /// @return Ethereum block that the most recent stake was created
-    function lastStakeBlock() external view returns (uint256);
+    function lastStakeBlock() external view returns (uint64);
 
     /// @return Number of active stakers currently staked
-    function stakerCount() external view returns (uint256);
+    function stakerCount() external view returns (uint64);
 
     /// @return Node hash as of this node number
-    function getNodeHash(uint256 index) external view returns (bytes32);
+    function getNodeHash(uint64 index) external view returns (bytes32);
 }
