+++ solgen/src/rollup/IRollupLogic.sol
@@ -29,7 +29,7 @@ interface IRollupUser {
 
     function requireUnresolvedExists() external view;
 
-    function countStakedZombies(uint256 nodeNum) external view returns (uint256);
+    function countStakedZombies(uint64 nodeNum) external view returns (uint256);
 }
 
 interface IRollupAdmin {
@@ -96,7 +96,7 @@ interface IRollupAdmin {
      * @notice Set number of blocks until a node is considered confirmed
      * @param newConfirmPeriod new number of blocks until a node is confirmed
      */
-    function setConfirmPeriodBlocks(uint256 newConfirmPeriod) external;
+    function setConfirmPeriodBlocks(uint64 newConfirmPeriod) external;
 
     /**
      * @notice Set the proving WASM module root
@@ -108,7 +108,7 @@ interface IRollupAdmin {
      * @notice Set number of extra blocks after a challenge
      * @param newExtraTimeBlocks new number of blocks
      */
-    function setExtraChallengeTimeBlocks(uint256 newExtraTimeBlocks) external;
+    function setExtraChallengeTimeBlocks(uint64 newExtraTimeBlocks) external;
 
     /**
      * @notice Set base stake required for an assertion
@@ -171,11 +171,11 @@ interface IRollupAdmin {
         uint256 inboxMaxCount,
         uint64 numBlocks,
         bool errored,
-        uint256 prevNode
+        uint64 prevNode
     ) external;
 
     function forceConfirmNode(
-        uint256 nodeNum,
+        uint64 nodeNum,
         bytes32 blockHash,
         bytes32 sendRoot
     ) external;
